Add TravisCI and update Kotlin to 1.1.4 (fix issue with kotlin package)
