+++ src/main/kotlin/kotlinext/jdbc/DataSources.kt
@@ -1,8 +1,7 @@
-package kotlin.jdbc
+package kotlinext.jdbc
 
 import java.sql.*
 import javax.sql.*
-import kotlin.template.StringTemplate
 
 /**
  * Processes a connection from the pool using the given function block
@@ -39,18 +38,4 @@ fun DataSource.update(sql: String): Int {
  */
 fun <T> DataSource.query(sql: String, block: (ResultSet) -> T): T {
     return use { it.query(sql, block) }
-}
-
-/**
- * Performs the update using the given SQL using a [[StringTemplate]]
- */
-fun DataSource.update(template : StringTemplate) : Int {
-    return use { it.update(template) }
-}
-
-/**
- * Perform a query on the connection using the SQL from the [[StringTemplate]] and processes the result set with a function
- */
-fun <T> DataSource.query(template : StringTemplate, resultBlock : (ResultSet) -> T) : T {
-    return use { it.query(template, resultBlock) }
 }
