+++ src/main/kotlin/kotlinext/jdbc/ResultSets.kt
@@ -1,4 +1,4 @@
-package kotlin.jdbc
+package kotlinext.jdbc
 
 import java.sql.*
